requirements.txt
pytest-django # Pytest
asgiref==3.8.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6 # Django
django-cors-headers==4.3.1 # CORS
django-filter==24.2 # Django Rest Framework
djangorestframework==3.15.1 # Django Rest Framework
djangorestframework-simplejwt==5.3.1 # JWT authentication plugin
pygments # code highlighting
markdown # Markdown support for the browsable API.
drf-nested-routers==0.93.5
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0 # basic tools for image processiong
pip==24.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
pytz==2024.1
requests==2.31.0
requests-unixsocket==0.3.0
setuptools==65.5.1
sqlparse==0.5.0
supervisor==4.2.5
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
gunicorn==20.1.0
psycopg2_binary # Postgres
python-dotenv==0.20.0
django-redis # redis with django
